[update] dm 코드 분리
